feat(plotting): flip summary_plot to prepare→spec→render (#638)
Extract SummaryPlotDataPreparer into cellpy.plotting.prepare.summary,
emit a real FigureSpec, and draw interactive figures through
PlotlyBackend.render. Delete PlotlyPlotBuilder and the provisional
CELLPY_SUMMARY_PLOTLY_SPEC dual path; keep SeabornPlotBuilder on the
same prepared frame until #639.
