+++ cellpy/plotting/__init__.py
@@ -14,16 +14,19 @@ live; the plan is `architecture-plan/cellpy2-plotting-redesign-plan.md`.
 | [`spec`][cellpy.plotting.spec] | ``FigureSpec`` / ``PanelSpec`` / ``AxisSpec`` |
 | [`registry`][cellpy.plotting.registry] | named ``PlotFamily`` records for ``summary_plot`` |
 | [`backends`][cellpy.plotting.backends] | render protocol + plotly formation layout (#637) |
+| [`prepare`][cellpy.plotting.prepare] | summary tidy-frame + ``FigureSpec`` (#638) |
+| [`context`][cellpy.plotting.context] | ``CellContext`` adapter (#638) |
 
 The old locations re-export from here, so nothing that imported them breaks.
-The drawing code itself — ``summary_plot`` and friends — still lives in
-``plotutils``; Stage 1 of epic #567 moves selection behind the registry and
-lands the spec/backends scaffolding for prepare→spec→render (#638).
+Public ``summary_plot`` still lives in ``plotutils`` but runs
+prepare → spec → ``PlotlyBackend.render`` for the interactive path (#638);
+``SeabornPlotBuilder`` remains until #639.
 """
 
 from __future__ import annotations
 
 from cellpy.plotting.backends import Backend, PlotlyBackend
+from cellpy.plotting.context import CellContext, from_source
 from cellpy.plotting.figures import (
     load_figure,
     load_matplotlib_figure,
@@ -32,6 +35,7 @@ from cellpy.plotting.figures import (
     save_matplotlib_figure,
 )
 from cellpy.plotting.labels import legend_replacer, remove_markers
+from cellpy.plotting.prepare import prepare_summary
 from cellpy.plotting.registry import (
     PlotFamily,
     _register_family,
@@ -44,12 +48,14 @@ from cellpy.plotting.theme import make_plotly_template
 __all__ = [
     "AxisSpec",
     "Backend",
+    "CellContext",
     "FigureSpec",
     "PanelSpec",
     "PlotFamily",
     "PlotlyBackend",
     "_register_family",
     "families",
+    "from_source",
     "get_family",
     "legend_replacer",
     "load_figure",
@@ -57,6 +63,7 @@ __all__ = [
     "load_plotly_figure",
     "make_matplotlib_manager",
     "make_plotly_template",
+    "prepare_summary",
     "remove_markers",
     "save_matplotlib_figure",
 ]
