+++ cellpy/plotting/backends/__init__.py
@@ -1,4 +1,4 @@
-"""Plotting backends (prepare → spec → render) — epic #567 / #637."""
+"""Plotting backends (prepare → spec → render) — epic #567 / #637–#638."""
 
 from __future__ import annotations
 
@@ -7,7 +7,6 @@ from cellpy.plotting.backends.plotly import (
     PlotlyBackend,
     configure_formation_layout,
     configure_fullcell_standard_domains,
-    use_spec_render,
 )
 
 __all__ = [
@@ -15,5 +14,4 @@ __all__ = [
     "PlotlyBackend",
     "configure_formation_layout",
     "configure_fullcell_standard_domains",
-    "use_spec_render",
 ]
