+++ cellpy/plotting/prepare/__init__.py
@@ -1,0 +1,7 @@
+"""Prepare stages for the prepare → spec → render pipeline (#638)."""
+
+from __future__ import annotations
+
+from cellpy.plotting.prepare.summary import prepare as prepare_summary
+
+__all__ = ["prepare_summary"]
