+++ cellpy/utils/plotutils.py
@@ -37,12 +37,6 @@ from cellpy.utils import helpers
 # Single copies of the plotting plumbing that used to be duplicated across
 # plotutils / collectors / batch_plotters (#567). Re-exported here so the
 # `from cellpy.utils.plotutils import load_figure` spelling keeps working.
-from cellpy.plotting.backends.plotly import (
-    DEFAULT_FORMATIONION_LABEL,
-    auto_range as _plotly_auto_range,
-    configure_formation_layout,
-    configure_fullcell_standard_domains,
-)
 from cellpy.plotting.figures import (  # noqa: F401
     load_figure,
     load_matplotlib_figure,
@@ -738,8 +732,8 @@ class SummaryPlotConfig:
         # block to draw, so ``show_formation`` must be False regardless of
         # how the caller set it. Without this, ``_mark_formation_cycles``
         # returns the ``slice(None, None)`` sentinel while ``show_formation``
-        # stays True, and ``_configure_formation_axes`` then evaluates
-        # ``~slice(...)`` which raises ``TypeError``. See issue #366.
+        # stays True, and formation layout then evaluates ``~slice(...)``
+        # which raises ``TypeError``. See issue #366.
         if self.formation_cycles is None:
             self.formation_cycles = 0
         self.formation_cycles = int(self.formation_cycles)
@@ -1005,854 +999,6 @@ class SummaryPlotInfo:
         self.y_trans = y_transformations
 
 
-class SummaryPlotDataPreparer:
-    """Handles data collection and transformation for summary plots.
-
-    Data preparation for summary_plot (extracted from the pre-2.0 monolith)
-    to improve maintainability and testability.
-    """
-
-    def __init__(self):
-        self.y_header = "value"
-        self.color = "variable"
-        self.row = "row"
-        self.col_id = "cycle_type"
-
-    def prepare_data(
-        self,
-        c: Any,
-        config: SummaryPlotConfig,
-        plot_info: SummaryPlotInfo,
-    ) -> dict:
-        """Prepare data for plotting.
-
-        Args:
-            c: cellpy object
-            config: SummaryPlotConfig with all parameters
-            summary_plot_info: SummaryPlotInfo containing information about pre-defined columns and labels
-        Returns:
-            Dictionary with keys:
-                - data: prepared DataFrame
-                - number_of_rows: number of rows for subplot layout
-                - x_label: x-axis label
-                - y_label: y-axis label
-                - max_cycle: maximum cycle number
-                - min_cycle: minimum cycle number
-                - max_val_normalized_col: max value for normalized columns
-                - formation_cycle_selector: boolean selector for formation cycles
-        """
-        x = config.x if config.x is not None else c.schema.summary.cycle_num
-        y = config.y
-        plot_registry.get(y)
-
-        number_of_rows = 1
-        max_val_normalized_col = 0.0
-
-        if config.hover_columns and (
-            y.startswith("fullcell_standard_")
-            or y.endswith("_split_constant_voltage")
-        ):
-            logging.warning(
-                "summary_plot: hover_columns is currently only supported for "
-                "standard plot types; ignoring for y=%r",
-                y,
-            )
-
-        # Prepare data based on plot type
-        if y.startswith("fullcell_standard_"):
-            s, number_of_rows = self._prepare_fullcell_standard_data(
-                c, x, y, plot_info.y_cols, plot_info.y_trans, config
-            )
-            max_val_normalized_col = (
-                s.loc[s["variable"].str.contains("retention"), "value"].max()
-                if len(s.loc[s["variable"].str.contains("retention")]) > 0
-                else 0.0
-            )
-        elif y.endswith("_split_constant_voltage"):
-            s, number_of_rows = self._prepare_cv_split_data(
-                c, x, y, plot_info.y_cols, config
-            )
-        else:
-            s, number_of_rows = self._prepare_standard_data(
-                c, x, y, plot_info.y_cols, config
-            )
-
-        # Calculate cycle ranges
-        max_cycle = s[x].max()
-        min_cycle = s[x].min()
-
-        # Get labels
-        x_label = plot_info.x_axis_labels.get(x, x)
-        if y in plot_info.y_axis_label:
-            y_label = plot_info.y_axis_label.get(y, y)
-        else:
-            y_label = y.replace("_", " ").title()
-
-        # Mark formation cycles
-        formation_cycle_selector = self._mark_formation_cycles(
-            s, x, config.formation_cycles, self.col_id
-        )
-
-        return {
-            "data": s,
-            "number_of_rows": number_of_rows,
-            "x_label": x_label,
-            "y_label": y_label,
-            "max_cycle": max_cycle,
-            "min_cycle": min_cycle,
-            "max_val_normalized_col": max_val_normalized_col,
-            "formation_cycle_selector": formation_cycle_selector,
-        }
-
-    def _prepare_fullcell_standard_data(
-        self, c, x, y, y_cols, y_trans, config
-    ) -> tuple:
-        """Prepare data for fullcell_standard plots."""
-
-        # The figure has 4 rows: coulombic efficiency, capacity, capacity retention, and CV capacity
-        number_of_rows = 4
-        column_set = y_cols[y]
-
-        summary = self._preprocess_summary(c, c.data.summary, config)
-        if summary.index.name == x:
-            summary = summary.reset_index(drop=False)
-
-        # Get CV-only summary
-        summary_only_cv = c.make_summary(
-            selector_type="only-cv", create_copy=True
-        ).data.summary
-        if summary_only_cv.index.name == x:
-            summary_only_cv = summary_only_cv.reset_index(drop=False)
-
-        # Merge summaries
-        s = summary.merge(summary_only_cv, on=x, how="outer", suffixes=("", "_cv"))
-        s = s.reset_index(drop=True)
-        s = s.melt(x)
-        s = s.loc[s.variable.isin(column_set)]
-
-        s[self.row] = 1  # default row for capacity
-
-        # Set row numbers using regex patterns
-        s.loc[s["variable"].str.contains(r"_efficiency$"), self.row] = (
-            0  # coulombic efficiency
-        )
-        s.loc[s["variable"].str.contains(r"cumulated.*loss"), self.row] = (
-            2  # cumulated loss
-        )
-        s.loc[s["variable"].str.startswith(r"mod_01_"), self.row] = (
-            2  # capacity retention
-        )
-        s.loc[s["variable"].str.contains(r"_cv$"), self.row] = 3  # cv data
-
-        # Reset losses if requested
-        if config.reset_losses:
-            logging.debug("Resetting losses")
-            first_values = (
-                s[s["variable"].str.contains(r"cumulated.*loss")]
-                .groupby("variable")["value"]
-                .transform("first")
-            )
-            mask = s["variable"].str.contains(r"cumulated.*loss")
-            s.loc[mask, "value"] = s.loc[mask, "value"] - first_values
-
-        # Apply normalization if requested
-        if config.fullcell_standard_normalization_type is not False:
-            logging.debug("Applying normalization")
-            s, max_val_normalized_col = self._apply_normalization(
-                s, y, y_trans, config, self.row
-            )
-
-        return s, number_of_rows
-
-    def _prepare_cv_split_data(self, c, x, y, y_cols, config) -> tuple:
-        """Prepare data for CV split plots."""
-        import warnings
-
-        if y.startswith("capacities_gravimetric"):
-            cap_type = "capacities_gravimetric"
-        elif y.startswith("capacities_areal"):
-            cap_type = "capacities_areal"
-        elif y.startswith("capacities_absolute"):
-            cap_type = "capacities_absolute"
-        else:
-            raise ValueError(f"Unknown capacity type for CV split: {y}")
-
-        column_set = y_cols[cap_type]
-
-        # Use partition_summary_cv_steps function
-        with warnings.catch_warnings():
-            warnings.simplefilter("ignore")
-            s = partition_summary_cv_steps(
-                c, x, column_set, config.split, self.color, self.y_header
-            )
-
-        number_of_rows = 3 if config.split else 1
-
-        return s, number_of_rows
-
-    def _prepare_standard_data(self, c, x, y, y_cols, config) -> tuple:
-        """Prepare data for standard plots."""
-        column_set = y_cols[y]
-        if isinstance(column_set, str):
-            column_set = [column_set]
-
-        summary = self._preprocess_summary(c, c.data.summary, config)
-        summary = summary.reset_index()
-
-        # Check if requested columns exist in summary
-        # For absolute capacities, fall back to base columns if _absolute columns don't exist
-        available_columns = set(summary.columns)
-        requested_columns = set(column_set)
-        missing_columns = requested_columns - available_columns
-
-        if missing_columns and y == "capacities_absolute":
-            # For absolute capacities, if _absolute columns don't exist, use base columns
-            hdr = c.headers_summary
-            base_columns = [hdr.charge_capacity_raw, hdr.discharge_capacity_raw]
-            # Check if base columns exist
-            if all(col in available_columns for col in base_columns):
-                column_set = base_columns
-            else:
-                # If base columns also don't exist, keep original column_set
-                # This will result in empty DataFrame, which will be handled downstream
-                pass
-        elif missing_columns:
-            # For other capacity types, if columns are missing, keep original column_set
-            # This will result in empty DataFrame, which will be handled downstream
-            pass
-
-        hover_cols = list(config.hover_columns or [])
-        if hover_cols:
-            missing = [h for h in hover_cols if h not in summary.columns]
-            if missing:
-                logging.warning(
-                    "summary_plot: dropping unknown hover_columns %s "
-                    "(available: %s)",
-                    missing,
-                    sorted(summary.columns),
-                )
-                hover_cols = [h for h in hover_cols if h in summary.columns]
-            # Avoid duplicating x and value columns in id_vars
-            hover_cols = [h for h in hover_cols if h != x and h not in column_set]
-
-        id_vars = [x, *hover_cols]
-        s = summary.melt(id_vars=id_vars)
-        s = s.loc[s.variable.isin(column_set)]
-        s = s.reset_index(drop=True)
-
-        # Check if we have any data after filtering
-        if len(s) == 0:
-            raise ValueError(
-                f"No data found for plot type '{y}'. "
-                f"Requested columns: {column_set}. "
-                f"Available columns in summary: {list(available_columns)}"
-            )
-
-        s[self.row] = 1
-
-        number_of_rows = 1
-        if config.split:
-            if y.endswith("_efficiency"):
-                s[self.row] = 1
-                s.loc[s["variable"].str.contains("efficiency"), self.row] = 0
-                number_of_rows = 2
-            elif y.endswith("_with_rate"):
-                hdr = c.headers_summary
-                rate_cols = {hdr.charge_c_rate, hdr.discharge_c_rate}
-                s[self.row] = 1
-                s.loc[s["variable"].isin(rate_cols), self.row] = 0
-                number_of_rows = 2
-
-        return s, number_of_rows
-
-    def _apply_normalization(self, s, y, y_trans, config, row_col) -> tuple:
-        """Apply normalization transformations to data."""
-        import re
-        from collections.abc import Iterable
-
-        max_val_normalized_col = 0.0
-        normalization_factor = config.fullcell_standard_normalization_factor
-        normalization_type = config.fullcell_standard_normalization_type
-        normalization_cycle_numbers = (
-            config.fullcell_standard_normalization_cycle_numbers
-        )
-
-        # TODO: check if this is really needed!!
-        # Determine normalization factor if not provided
-        if normalization_factor is None:
-            logging.debug(
-                f"No normalization factor provided for {y}, using {normalization_type}"
-            )
-
-        if y.startswith("fullcell_standard_cumloss_") and normalization_type != "max":
-            logging.debug("only allowing for 'max' for cumloss plots")
-            normalization_type = "max"
-
-        if normalization_type in ["on-cycles", "on-cycle"]:
-            if normalization_cycle_numbers is None:
-                raise ValueError(
-                    "Normalization cycle numbers are required for on-cycles normalization"
-                )
-            if isinstance(normalization_cycle_numbers, Iterable):
-                cycle_numbers = [cycle - 1 for cycle in normalization_cycle_numbers]
-            else:
-                cycle_numbers = [normalization_cycle_numbers - 1]
-            normalization_cycle_numbers = cycle_numbers
-
-        trans_kwargs = dict(
-            normalization_factor=normalization_factor,
-            normalization_type=normalization_type,
-            normalization_scaler=config.fullcell_standard_normalization_scaler,
-            normalization_indexes=normalization_cycle_numbers,
-        )
-
-        # Transform the data
-        max_row_val = s[row_col].max()
-        for col, trans_dict in y_trans.get(y, {}).items():
-            for (new_row_val, new_col), trans in trans_dict.items():
-                if new_col in s["variable"].values:
-                    # Transforming on existing column
-                    s.loc[s["variable"] == col, "value"] = trans(
-                        s.loc[s["variable"] == col, "value"].values, **trans_kwargs
-                    )
-                else:
-                    # Creating new column
-                    old_col = col
-                    if new_row_val is not None:
-                        row_val = new_row_val
-                    else:
-                        row_val = s.loc[s["variable"] == col, row_col]
-                        if not row_val.empty:
-                            row_val = row_val.values[0]
-                        else:
-                            max_row_val += 1
-                            row_val = max_row_val
-
-                    if old_col.startswith("mod_"):
-                        old_col = re.sub(r"^mod_\d{2}_", "", old_col)
-                    new_col_frame_section = s.loc[s["variable"] == old_col].copy()
-                    new_col_frame_section["variable"] = new_col
-                    new_col_frame_section[row_col] = row_val
-                    transformed_values = trans(
-                        new_col_frame_section["value"].values, **trans_kwargs
-                    )
-                    new_col_frame_section["value"] = transformed_values
-                    s = pd.concat([s, new_col_frame_section], ignore_index=True)
-                    s = s.reset_index(drop=True)
-                    s = s.sort_values(by=[row_col, "variable"])
-
-                max_val_normalized_col = s.loc[s["variable"] == new_col, "value"].max()
-
-        return s, max_val_normalized_col
-
-    def _mark_formation_cycles(self, s, x, formation_cycles, col_id):
-        """Mark formation cycles in the data."""
-        formation_cycle_selector = slice(None, None)
-        if formation_cycles > 0:
-            formation_cycle_selector = s[x] <= formation_cycles
-            s[col_id] = "standard"
-            s.loc[formation_cycle_selector, col_id] = "formation"
-        return formation_cycle_selector
-
-    @staticmethod
-    def _preprocess_summary(c: Any, summary: pd.DataFrame, config) -> pd.DataFrame:
-        """Apply optional rate-rescaling and row filtering to a summary copy.
-
-        Two opt-in steps, both no-ops when their config field is ``None``:
-
-        * ``config.nominal_capacity`` rescales the existing
-          ``charge_c_rate`` / ``discharge_c_rate`` columns to use a new
-          nominal capacity instead of the one set on the cell. The
-          rescale factor is ``c.data.nom_cap / nominal_capacity``: since
-          ``rate = current / nom_cap``, the rate columns are multiplied
-          by ``old_nom_cap / new_nom_cap``.
-        * ``config.filters`` is forwarded to
-          :func:`cellpy.filters.filter_summary`. The default
-          ``rate_filter_columns`` resolves to both rate columns from
-          ``c.headers_summary`` (charge AND discharge).
-
-        Operates on a copy; the caller's ``summary`` argument is not
-        mutated.
-        """
-        out = summary.copy() if summary is not None else summary
-
-        if config.nominal_capacity is not None:
-            hdr = c.headers_summary
-            old_nom_cap = getattr(c.data, "nom_cap", None)
-            if old_nom_cap in (None, 0):
-                logging.warning(
-                    "summary_plot: nominal_capacity override requested but "
-                    "cell.data.nom_cap is %r; skipping rate rescale.",
-                    old_nom_cap,
-                )
-            else:
-                scale = float(old_nom_cap) / float(config.nominal_capacity)
-                for col in (hdr.charge_c_rate, hdr.discharge_c_rate):
-                    if col in out.columns:
-                        out[col] = out[col] * scale
-                logging.debug(
-                    "summary_plot: rescaled rate columns by %.6g "
-                    "(old nom_cap=%s, new=%s)",
-                    scale,
-                    old_nom_cap,
-                    config.nominal_capacity,
-                )
-
-        if config.filters:
-            from cellpy.filters import filter_summary
-
-            hdr = c.headers_summary
-            filter_kwargs = dict(config.filters)
-            if (
-                "rate" in filter_kwargs
-                and "rate_columns" not in filter_kwargs
-            ):
-                if config.rate_filter_columns is not None:
-                    filter_kwargs["rate_columns"] = config.rate_filter_columns
-                else:
-                    filter_kwargs["rate_columns"] = (
-                        hdr.charge_c_rate,
-                        hdr.discharge_c_rate,
-                    )
-            before = len(out)
-            out = filter_summary(out, **filter_kwargs)
-            logging.debug(
-                "summary_plot: filters %s reduced rows %d -> %d",
-                filter_kwargs,
-                before,
-                len(out),
-            )
-
-        return out
-
-
-class PlotlyPlotBuilder:
-    """Handles Plotly-specific plotting logic for summary plots.
-
-    Plotly rendering for summary_plot (extracted from the pre-2.0 monolith)
-    to improve maintainability and testability.
-    """
-
-    def __init__(self):
-        self.y_header = "value"
-        self.color = "variable"
-        self.row = "row"
-        self.col_id = "cycle_type"
-
-    def build_plot(
-        self,
-        data: pd.DataFrame,
-        prepared_data_info: dict,
-        config: SummaryPlotConfig,
-        additional_kwargs: dict,
-        c: Any,
-    ) -> Any:
-        """Build Plotly figure from prepared data.
-
-        Args:
-            data: Prepared DataFrame from SummaryPlotDataPreparer
-            prepared_data_info: Dictionary with metadata from data preparer
-            config: SummaryPlotConfig with all parameters
-            additional_kwargs: Additional kwargs for plotly (from legacy function)
-            c: cellpy object (needed for some label generation)
-
-        Returns:
-            Plotly figure object
-        """
-        import plotly.express as px
-
-        # Extract plotly-specific parameters from additional_kwargs
-        smart_link = additional_kwargs.pop("smart_link", True)
-        show_y_labels_on_right_pane = additional_kwargs.pop(
-            "show_y_labels_on_right_pane", False
-        )
-        plotly_row_ratios = additional_kwargs.pop(
-            "fullcell_standard_row_height_ratios", [0.3, 0.6, 0.9]
-        )
-        plotly_row_space = additional_kwargs.pop("fullcell_standard_row_space", 0.02)
-
-        # Extract plotly_* parameters for update_traces
-        plotly_update_traces = {}
-        for k in list(additional_kwargs.keys()):
-            if k.startswith("plotly_"):
-                plotly_update_traces[k.replace("plotly_", "")] = additional_kwargs.pop(
-                    k
-                )
-
-        # Set default title if not provided
-        title = config.title
-        if title is None:
-            title = f"Summary <b>{c.cell_name}</b>"
-
-        x = config.x if config.x is not None else c.schema.summary.cycle_num
-        y = config.y
-        number_of_rows = prepared_data_info["number_of_rows"]
-        x_label = prepared_data_info["x_label"]
-        y_label = prepared_data_info["y_label"]
-        max_cycle = prepared_data_info["max_cycle"]
-        min_cycle = prepared_data_info["min_cycle"]
-        max_val_normalized_col = prepared_data_info["max_val_normalized_col"]
-        formation_cycle_selector = prepared_data_info["formation_cycle_selector"]
-
-        # Prepare plotly kwargs
-        plotly_kwargs = {
-            "color": self.color,
-            "height": config.height,
-            "markers": config.markers,
-            "title": title,
-            "width": config.width,
-        }
-
-        # Add facet_row if split
-        if config.split and self.row in data.columns:
-            plotly_kwargs["facet_row"] = self.row
-
-        # Add hover columns if they survived data preparation
-        if config.hover_columns:
-            present = [h for h in config.hover_columns if h in data.columns]
-            if present:
-                plotly_kwargs["hover_data"] = present
-
-        # Set default height if not provided
-        if plotly_kwargs.get("height") is None:
-            if y.startswith("fullcell_standard_"):
-                plotly_kwargs["height"] = 800
-            elif config.split and number_of_rows > 1:
-                plotly_kwargs["height"] = 800
-            else:
-                plotly_kwargs["height"] = 200 + 200 * number_of_rows
-
-        # Set plotly template
-        set_plotly_template(config.plotly_template)
-
-        # Add facet_col for formation cycles
-        if config.show_formation and self.col_id in data.columns:
-            plotly_kwargs["facet_col"] = self.col_id
-
-        # Create base figure
-
-        fig = px.line(
-            data,
-            x=x,
-            y=self.y_header,
-            **plotly_kwargs,
-            labels={
-                x: x_label,
-                self.y_header: y_label,
-            },
-            **additional_kwargs,
-        )
-
-        # Update traces
-        if plotly_update_traces:
-            fig.update_traces(**plotly_update_traces)
-
-        # Hide legend if requested
-        if not config.show_legend:
-            fig.update_layout(showlegend=False)
-
-        # Apply y_range if provided
-        if config.y_range is not None:
-            fig.update_layout(yaxis=dict(range=config.y_range))
-
-        # Configure formation cycles and subplot layouts
-        if config.show_formation:
-            self._configure_formation_axes(
-                fig,
-                data,
-                x,
-                config,
-                number_of_rows,
-                max_cycle,
-                min_cycle,
-                formation_cycle_selector,
-                show_y_labels_on_right_pane,
-                y,
-                max_val_normalized_col,
-                plotly_row_ratios,
-                plotly_row_space,
-                c,
-            )
-        else:
-            # Configure without formation cycles
-            self._configure_no_formation_axes(
-                fig,
-                config,
-                y,
-                number_of_rows,
-                max_val_normalized_col,
-                plotly_row_ratios,
-                plotly_row_space,
-                c,
-            )
-
-        # Apply x_range if provided
-        if config.x_range is not None:
-            if not config.show_formation:
-                fig.update_layout(xaxis=dict(range=config.x_range))
-
-        # Handle split and share_y
-        if config.split:
-            if config.show_formation:
-                if not config.share_y and not smart_link:
-                    fig.update_yaxes(matches=None)
-            elif not config.share_y:
-                fig.update_yaxes(matches=None)
-
-        # Add rangeslider if requested
-        if config.rangeslider:
-            if config.show_formation:
-                logging.critical(
-                    "Can not add rangeslider when showing formation cycles"
-                )
-            else:
-                fig.update_layout(xaxis_rangeslider_visible=True)
-
-        # Auto-convert legend labels
-        if config.auto_convert_legend_labels and config.show_legend:
-            self._convert_legend_labels(fig)
-
-        return fig
-
-    def _auto_range(self, fig: Any, axis_name_1: str, axis_name_2: str) -> list:
-        """Calculate auto range for two y-axes (only works for plotly)."""
-        return _plotly_auto_range(fig, axis_name_1, axis_name_2)
-
-    def _configure_formation_axes(
-        self,
-        fig,
-        data,
-        x,
-        config,
-        number_of_rows,
-        max_cycle,
-        min_cycle,
-        formation_cycle_selector,
-        show_y_labels_on_right_pane,
-        y,
-        max_val_normalized_col,
-        plotly_row_ratios,
-        plotly_row_space,
-        c,
-    ):
-        """Configure axes when showing formation cycles.
-
-        Thin adapter over :func:`cellpy.plotting.backends.plotly.configure_formation_layout`
-        (#637). Per-row-count methods are gone; the N-row grid lives in the backend.
-        """
-        x_axis_domain_formation = [
-            0.0,
-            config.x_axis_domain_formation_fraction - config.column_separator / 2,
-        ]
-        x_axis_domain_rest = [
-            config.x_axis_domain_formation_fraction + config.column_separator / 2,
-            0.95,
-        ]
-        max_cycle_formation = data.loc[formation_cycle_selector, x].max()
-        min_cycle_rest = data.loc[~formation_cycle_selector, x].min()
-
-        if x == _NORMALIZED_CYCLE_INDEX:
-            dd = 0.1
-        else:
-            dd = 0.4
-        x_axis_range_formation = [min_cycle - dd, max_cycle_formation + dd]
-        x_axis_range_rest = [min_cycle_rest - dd, max_cycle + dd]
-
-        if config.x_range is not None:
-            x_axis_range_rest = [
-                x_axis_range_rest[0],
-                min(config.x_range[1], x_axis_range_rest[1]),
-            ]
-
-        eff_lim = config.ce_range
-        top_row_label = _plotly_top_row_label(y) if number_of_rows == 2 else None
-
-        # Pre-resolve per-row y ranges that the old per-N methods special-cased.
-        # ``None`` entries mean "auto-range that facet-row pair".
-        row_y_ranges: list[Optional[list]] = [None] * number_of_rows
-        if number_of_rows == 2:
-            row_y_ranges[0] = config.y_range
-            row_y_ranges[1] = eff_lim
-        elif number_of_rows == 4 and y.startswith("fullcell_standard_"):
-            # Row order matches the old 4-row method: 0=CV, 1=retention/norm,
-            # 2=capacity, 3=CE. ``None`` → auto-range inside the layout helper.
-            row_y_ranges[0] = config.cv_share_range
-            if config.fullcell_standard_normalization_type is not False:
-                row_y_ranges[1] = config.norm_range or [
-                    0.0,
-                    max(
-                        max_val_normalized_col,
-                        config.fullcell_standard_normalization_scaler,
-                    ),
-                ]
-            row_y_ranges[2] = config.y_range
-            row_y_ranges[3] = config.ce_range
-
-        configure_formation_layout(
-            fig,
-            n_rows=number_of_rows,
-            x_axis_domain_formation=x_axis_domain_formation,
-            x_axis_domain_rest=x_axis_domain_rest,
-            x_axis_range_formation=x_axis_range_formation,
-            x_axis_range_rest=x_axis_range_rest,
-            show_y_labels_on_right_pane=show_y_labels_on_right_pane,
-            formation_header=DEFAULT_FORMATIONION_LABEL,
-            row_y_ranges=row_y_ranges,
-            top_row_label=top_row_label,
-        )
-
-        if number_of_rows == 4 and y.startswith("fullcell_standard_"):
-            self._configure_fullcell_standard_domains(
-                fig,
-                config,
-                plotly_row_ratios,
-                plotly_row_space,
-                c,
-                y,
-            )
-
-    def _configure_fullcell_standard_domains(
-        self,
-        fig,
-        config,
-        plotly_row_ratios,
-        plotly_row_space,
-        c,
-        y,
-    ):
-        """Configure domain layout for fullcell_standard plots."""
-        mode = y.split("_")[-1]
-        configure_fullcell_standard_domains(
-            fig,
-            plotly_row_ratios=plotly_row_ratios,
-            plotly_row_space=plotly_row_space,
-            capacity_unit=_get_capacity_unit(c, mode=mode),
-            y=y,
-            show_formation=config.show_formation,
-            x_axis_domain_formation_fraction=config.x_axis_domain_formation_fraction,
-            link_capacity_scales=config.link_capacity_scales,
-            normalization_type=config.fullcell_standard_normalization_type,
-            normalization_factor=config.fullcell_standard_normalization_factor,
-            normalization_scaler=config.fullcell_standard_normalization_scaler,
-        )
-
-    def _configure_no_formation_axes(
-        self,
-        fig,
-        config,
-        y,
-        number_of_rows,
-        max_val_normalized_col,
-        plotly_row_ratios,
-        plotly_row_space,
-        c,
-    ):
-        """Configure axes when not showing formation cycles."""
-        eff_lim = config.ce_range
-
-        _top_label = _plotly_top_row_label(y)
-        if _top_label is not None:
-            fig.update_layout(
-                yaxis=dict(domain=[0.0, 0.65]),
-                yaxis2={
-                    "title": dict(text=_top_label),
-                    "domain": [0.7, 1.0],
-                },
-            )
-        if y.startswith("fullcell_standard_"):
-            range_1 = eff_lim or self._auto_range(fig, "y4", "y4")
-            range_2 = config.y_range or self._auto_range(fig, "y3", "y3")
-            range_3 = self._auto_range(fig, "y2", "y2")
-            if config.fullcell_standard_normalization_type is not False:
-                range_3 = [
-                    0.0,
-                    max(
-                        max_val_normalized_col,
-                        config.fullcell_standard_normalization_scaler,
-                    ),
-                ]
-            range_3 = config.norm_range or range_3
-
-            range_4 = config.cv_share_range or self._auto_range(fig, "y", "y")
-            fig.layout["annotations"] = 4 * [PLOTLY_BLANK_LABEL]
-
-            ce_domain_start, ce_domain_end = plotly_row_ratios[2], 1.0
-            capacity_domain_start, capacity_domain_end = (
-                plotly_row_ratios[1],
-                plotly_row_ratios[2] - plotly_row_space,
-            )
-            loss_domain_start, loss_domain_end = (
-                plotly_row_ratios[0],
-                plotly_row_ratios[1] - plotly_row_space,
-            )
-            cv_domain_start, cv_domain_end = (
-                0.0,
-                plotly_row_ratios[0] - plotly_row_space,
-            )
-
-            # Format y-axis labels with HTML for proper alignment
-            capacity_unit = _get_capacity_unit(c, mode=y.split("_")[-1])
-            ce_label = "Coulombic<br>Efficiency (%)"
-            capacity_label = f"Capacity<br>({capacity_unit})"
-            if (
-                config.fullcell_standard_normalization_type
-                and config.fullcell_standard_normalization_factor is not None
-            ):
-                _norm_label = f"[{config.fullcell_standard_normalization_scaler:.1f}/{config.fullcell_standard_normalization_factor:.1f} {capacity_unit}]"
-                loss_label = f"Capacity<br>Retention (norm.)<br>{_norm_label}"
-            else:
-                loss_label = f"Capacity<br>Retention ({capacity_unit})"
-            cv_label = f"CV Capacity<br>({capacity_unit})"
-
-            fig.update_layout(
-                yaxis4={
-                    "title": dict(text=ce_label),
-                    "domain": [ce_domain_start, ce_domain_end],
-                    "matches": None,
-                    "range": range_1,
-                },
-                yaxis3={
-                    "title": dict(text=capacity_label),
-                    "domain": [capacity_domain_start, capacity_domain_end],
-                    "matches": None,
-                    "range": range_2,
-                },
-                yaxis2={
-                    "title": dict(text=loss_label),
-                    "domain": [loss_domain_start, loss_domain_end],
-                    "matches": None,
-                    "range": range_3,
-                },
-                yaxis={
-                    "title": dict(text=cv_label),
-                    "domain": [cv_domain_start, cv_domain_end],
-                    "matches": None,
-                    "range": range_4,
-                },
-            )
-
-    def _convert_legend_labels(self, fig):
-        """Convert legend labels to nicer format."""
-        for trace in fig.data:
-            name = trace.name
-            name = name.replace("_", " ").title()
-            name = name.replace("Gravimetric", "Grav.")
-            name = name.replace("Cv", "(CV)")
-            name = name.replace("Non (CV)", "(without CV)")
-            hover_template = trace.hovertemplate
-            if hover_template:
-                statements = []
-                for statement in hover_template.split("<br>"):
-                    if "=" in statement:
-                        variable, value = statement.split("=", 1)
-                        if value.startswith("%{y}"):
-                            variable = name
-                        statement = "=".join((variable, value))
-                    statements.append(statement)
-                hover_template = "<br>".join(statements)
-            trace.update(name=name, hovertemplate=hover_template)
-
-
 class SeabornPlotBuilder:
     """Handles Seaborn-specific plotting logic for summary plots.
 
@@ -1877,8 +1023,8 @@ class SeabornPlotBuilder:
         """Build Seaborn/Matplotlib figure from prepared data.
 
         Args:
-            data: Prepared DataFrame from SummaryPlotDataPreparer
-            prepared_data_info: Dictionary with metadata from data preparer
+            data: Prepared DataFrame from ``cellpy.plotting.prepare.summary``
+            prepared_data_info: Dictionary with metadata from prepare
             config: SummaryPlotConfig with all parameters
             additional_kwargs: Additional kwargs for seaborn (from legacy function)
             c: cellpy object (needed for some label generation)
@@ -3041,27 +2187,34 @@ def summary_plot(
             )
             return None
 
-    # Prepare data
+    # prepare → spec → render (#638)
+    from cellpy.plotting.backends.plotly import PlotlyBackend
+    from cellpy.plotting.context import from_source
+    from cellpy.plotting.prepare.summary import prepare as prepare_summary
+
     plot_info = SummaryPlotInfo(c)
-    preparer = SummaryPlotDataPreparer()
-    prepared_data_info = preparer.prepare_data(
-        c,
-        config,
-        plot_info,
-    )
+    family = plot_registry.get(config.y)
+    ctx = from_source(c)
+    frame, spec = prepare_summary(ctx, family, config, plot_info=plot_info)
 
-    builder = PlotlyPlotBuilder() if config.interactive else SeabornPlotBuilder()
-
-    fig = builder.build_plot(
-        prepared_data_info["data"],
-        prepared_data_info,
-        config,
-        config.additional_kwargs,
-        c,
-    )
+    if config.interactive:
+        fig = PlotlyBackend().render(frame, spec)
+    else:
+        prepared_data_info = dict(spec.extras.get("prepared_data_info") or {})
+        # Seaborn builder still expects the frame under the old info dict shape.
+        prepared_data_info.setdefault("number_of_rows", spec.extras.get("number_of_rows"))
+        prepared_data_info.setdefault("x_label", spec.x_axis.label)
+        prepared_data_info.setdefault("y_label", spec.extras.get("y_label"))
+        fig = SeabornPlotBuilder().build_plot(
+            frame,
+            prepared_data_info,
+            config,
+            dict(config.additional_kwargs),
+            c,
+        )
 
     if config.return_data:
-        return fig, prepared_data_info["data"]
+        return fig, frame
     return fig
